older torch support, other fixes
